use ServerConfig::into_server instead of server::from_config
